+++ src/repo/entities/repo.entity.ts
@@ -552,4 +552,28 @@ export class DbRepo extends BaseEntity {
     insert: false,
   })
   public pr_active_count?: number;
+
+  @ApiModelProperty({
+    description: "Number of commits and comments over the number of unique contributors in time range",
+    example: 8.82,
+    type: "float",
+  })
+  @Column({
+    type: "float",
+    select: false,
+    insert: false,
+  })
+  public activity_ratio?: number;
+
+  @ApiModelProperty({
+    description: "Health is an aggregate, overview calculation of how a repo is generally doing",
+    example: 1.23,
+    type: "float",
+  })
+  @Column({
+    type: "float",
+    select: false,
+    insert: false,
+  })
+  public health?: number;
 }