+++ src/workspace/dtos/update-workspace-contributors.dto.ts
@@ -10,9 +10,9 @@ export class UpdateWorkspaceContributorsDto {
   @ApiProperty({
     description: "An array of contributor objects to be added to the workspace",
     isArray: true,
-    example: [{ id: 12345 }],
+    example: [{ id: 12345, login: "jpmcb" }],
   })
   @IsArray()
   @Type(() => NewWorkspaceContributor)
-  contributors: { id: number }[];
+  contributors: { id?: number; login?: string }[];
 }